Refactor location bar into 2 separate components
